+++ polaris-react/src/components/Pagination/Pagination.stories.tsx
@@ -57,3 +57,27 @@ export function WithLabel() {
     />
   );
 }
+
+export function WithTableType() {
+  return (
+    <div
+      style={{
+        maxWidth: 'calc(700px + (2 * var(--p-space-4)))',
+        margin: '0 auto',
+        border: '1px solid var(--p-color-border)',
+      }}
+    >
+      <Pagination
+        onPrevious={() => {
+          console.log('Previous');
+        }}
+        onNext={() => {
+          console.log('Next');
+        }}
+        type="table"
+        hasNext
+        label="1-50 of 8,450 orders"
+      />
+    </div>
+  );
+}