+++ polaris-react/src/components/Pagination/Pagination.tsx
@@ -11,6 +11,8 @@ import {Text} from '../Text';
 import {Tooltip} from '../Tooltip';
 import {Box} from '../Box';
 import {useFeatures} from '../../utilities/features';
+import {HorizontalStack} from '../HorizontalStack';
+import {classNames} from '../../utilities/css';
 
 import styles from './Pagination.scss';
 
@@ -46,6 +48,8 @@ export interface PaginationProps {
   onPrevious?(): void;
   /** Text to provide more context in between the arrow buttons */
   label?: React.ReactNode;
+  /** Layout structure of the component */
+  type?: 'page' | 'table';
 }
 
 export function Pagination({
@@ -62,6 +66,7 @@ export function Pagination({
   accessibilityLabel,
   accessibilityLabels,
   label,
+  type = 'page',
 }: PaginationProps) {
   const i18n = useI18n();
   const {polarisSummerEditions2023} = useFeatures();
@@ -159,6 +164,43 @@ export function Pagination({
         }
       />
     ));
+
+  if (type === 'table') {
+    const labelMarkup = label ? (
+      <Text as="span" variant="bodySm" fontWeight="medium">
+        {label}
+      </Text>
+    ) : null;
+
+    return (
+      <nav
+        aria-label={navLabel}
+        ref={node}
+        className={classNames(styles.Pagination, styles.table)}
+      >
+        {previousButtonEvents}
+        {nextButtonEvents}
+        <Box
+          background="bg-subdued"
+          paddingBlockStart="1_5-experimental"
+          paddingBlockEnd="1_5-experimental"
+          paddingInlineStart="3"
+          paddingInlineEnd="2"
+        >
+          <HorizontalStack
+            align={labelMarkup ? 'space-between' : 'end'}
+            blockAlign="center"
+          >
+            {labelMarkup}
+            <ButtonGroup segmented>
+              {constructedPrevious}
+              {constructedNext}
+            </ButtonGroup>
+          </HorizontalStack>
+        </Box>
+      </nav>
+    );
+  }
 
   const labelTextMarkup =
     hasNext && hasPrevious ? (