+++ polaris-react/src/components/TextField/TextField.stories.tsx
@@ -866,3 +866,19 @@ export function WithFormSubmit() {
     </VerticalStack>
   );
 }
+
+export function With1PasswordDisabled() {
+  const [value, setValue] = useState('Jaded Pixel');
+
+  const handleChange = useCallback((newValue) => setValue(newValue), []);
+
+  return (
+    <TextField
+      label="Store name"
+      value={value}
+      onChange={handleChange}
+      autoComplete="off"
+      disable1Password
+    />
+  );
+}